fix usage of enum.keys()
A named enum acts like a Dictionary, so .keys() returns
an Array of the string keys, not a Dictionary. This happens
to work if enum values start at 0 and are contiguous but not
otherwise. Use find_key even if it's not efficient since the
enum-to-string lookup is only used for debugging.